+++ src/config/constants.ts
@@ -35,8 +35,7 @@ export const STORAGE_KEYS = {
     DEPARTURE_SORT: 'departureSort',
     LAST_LOCATION: 'lastUserLocation',
     WELCOME_SEEN: 'departs_welcome_seen',
-    FAVORITES: 'favoriteStops',
-    SHOW_HEATMAP: 'showHeatmap'
+    FAVORITES: 'favoriteStops'
 };
 
 /**