+++ src/contexts/MapContext.tsx
@@ -71,8 +71,7 @@ export const MapProvider: React.FC<{ children: React.ReactNode; mapRef: React.Re
         setRouteFilter,
         setBounds,
         setDebouncedBounds,
-        toggleFavorite,
-        setShowHeatmap
+        toggleFavorite
     } = useMapReducer();
 
     const { userLocation, isGeoPending, handleLocate, performGeolocation } = useGeolocation(mapRef);
@@ -231,8 +230,7 @@ export const MapProvider: React.FC<{ children: React.ReactNode; mapRef: React.Re
             handleDepartureClick,
             performGeolocation,
             setMapLoaded,
-            setLabelLayerId,
-            setShowHeatmap
+            setLabelLayerId
         },
         mapEvents: {
             onMove,